requirements.txt
openai>=0.27.0
python-dotenv>=0.19.0
tqdm>=4.60.0
requests>=2.31.0